Attempts to use Spotify's KJU Remoting to set coverart and progress
It does in fact update Spotify's audioHmiState values! It changes the
side Media panel from the Artist/Song multimediaInfoEvent to instead show a black square
with empty metadata, exactly like Spotify clears the view in #93
If the Spotify app is then selected, while Spotify is not trying to update
it's own audioHmiState, the coverart and progress then pop into the side bar.
The artist and title shows up in the audioHmiState fields, and the
progress bar updates during playback, and coverart changes on song changes.
However, switching back to Audioplayer for context clears it back to black square.
So, while promising, it's probably not a solution for #59